skills/ig-transcript/requirements.txt
yt-dlp>=2024.1.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0